% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmpost.R
\name{dfmpost}
\alias{dfmpost}
\title{Posterior Simulation for Dynamic Factor Models}
\usage{
dfmpost(object)
}
\arguments{
\item{object}{an object of class 'dfmodel', usually, a result of a
call to \code{\link{create_dfmodel}} in combination with \code{\link{add_priors.dfmodel}}
and \code{\link{add_initial_values.dfmodel}}.}
}
\value{
An object of class 'dfm'.
}
\description{
Produces draws from the posterior distributions of Bayesian dynamic factor models.
}
\details{
The function implements the posterior simulation algorithm for Bayesian dynamic factor models.

The implementation follows the description in Chan et al. (2019).
}
\examples{

# Load data
data("bem_dfmdata")

# Generate model data
model <- create_dfmodel(x = bem_dfmdata, p = 1, n = 1,
                        iterations = 20, burnin = 10)
# Number of iterations and burnin should be much higher.

# Add prior specifications
model <- add_priors(model,
                    lambda = list(vinv = .01),
                    u = list(shape = 5, rate = 4),
                    a = list(vinv = .01),
                    v = list(shape = 5, rate = 4))

# Add initial values
model <- add_initial_values(model)

# Obtain posterior draws
object <- dfmpost(model)

}
\references{
Chan, J., Koop, G., Poirier, D. J., & Tobias J. L. (2019). \emph{Bayesian econometric methods}
(2nd ed.). Cambridge: Cambridge University Press.
}
