% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_forecast_errors.R
\name{get_forecast_errors.dfmodel}
\alias{get_forecast_errors.dfmodel}
\alias{get_forecast_errors.favarmodel}
\title{Get Forecast Errors of a Factor Model}
\usage{
\method{get_forecast_errors}{dfmodel}(object, ...)

\method{get_forecast_errors}{favarmodel}(object, ...)
}
\arguments{
\item{object}{an object of class \code{'dfmodel'} or \code{'favarmodel'}.}

\item{...}{not used.}
}
\value{
The \code{\link[coda]{mcmc}} object in \code{posterior$forecast$errors}, or
\code{NULL} where there is none.
}
\description{
Returns the forecast errors \code{\link{add_forecast_errors.dfmodel}} added to a model.
}
