% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.dfm.R
\name{predict.dfm}
\alias{predict.dfm}
\title{Predict Method for Objects of Class dfm}
\usage{
\method{predict}{dfm}(object, n_ahead = 10, ci = 0.95, ...)
}
\arguments{
\item{object}{an object of class 'dfm'.}

\item{n_ahead}{number of steps ahead at which to predict.}

\item{ci}{a numeric between 0 and 1 specifying the probability mass covered by the
credible intervals. Defaults to 0.95.}

\item{...}{additional arguments.}
}
\description{
Forecasting method for class 'dfm' with credible bands.
}
