% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_posterior_coefficients.favarmodel.R
\name{add_posterior_coefficients.favarmodel}
\alias{add_posterior_coefficients.favarmodel}
\title{Estimate a Factor Augmented VAR}
\usage{
\method{add_posterior_coefficients}{favarmodel}(object, posterior_function = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{'favarmodel'}.}

\item{posterior_function}{a function that estimates the model. Used to override the built-in sampler.}

\item{...}{not used.}
}
\value{
The model object with the result attached under \code{posterior}.
}
\description{
Runs the Gibbs sampler and attaches the posterior draws.
}
