% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_posterior_loglik.favarmodel.R
\name{add_posterior_loglik.favarmodel}
\alias{add_posterior_loglik.favarmodel}
\title{Log Likelihood of a Factor Augmented VAR}
\usage{
\method{add_posterior_loglik}{favarmodel}(object, ...)
}
\arguments{
\item{object}{an object of class \code{'favarmodel'}.}

\item{...}{not used.}
}
\value{
The model object with the result attached under \code{posterior}.
}
\description{
Computes the pointwise log likelihood, draws by periods.
}
