% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmtools-package.R
\docType{package}
\name{dfmtools-package}
\alias{dfmtools}
\alias{dfmtools-package}
\title{dfmtools: Bayesian Inference of Dynamic Factor Models}
\description{
Assists in the set-up of algorithms for Bayesian inference of dynamic factor models and factor augmented vector autoregressive models. Functions for posterior simulation and forecasting are largely based on the introductory texts of Chan, Koop, Poirier and Tobias (2019, ISBN: 9781108437493) and Luetkepohl (2006, ISBN: 9783540262398).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/franzmohr/dfmtools}
  \item Report bugs at \url{https://github.com/franzmohr/dfmtools/issues}
}

}
\author{
\strong{Maintainer}: Franz X. Mohr \email{franz.x.mohr@outlook.com} (\href{https://orcid.org/0009-0003-8890-7781}{ORCID})

Authors:
\itemize{
  \item Franz X. Mohr \email{franz.x.mohr@outlook.com} (\href{https://orcid.org/0009-0003-8890-7781}{ORCID})
}

}
\keyword{internal}
