% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_posterior.dfmodel.R
\name{draw_posterior.dfmodel}
\alias{draw_posterior.dfmodel}
\title{Posterior Simulation}
\usage{
\method{draw_posterior}{dfmodel}(object, ...)
}
\arguments{
\item{object}{a list of model specifications, which should be passed on
to function \code{FUN}. Usually, the output of a call to \code{\link{create_dfmodel}}
in combination with \code{\link{add_priors.dfmodel}} and \code{\link{add_initial_values.dfmodel}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of the class of the output of the applied posterior
simulation function. In case the package's own function is used, this will
result in an object of class 'dfm'.
}
\description{
Forwards model input to posterior simulation functions.
}
\examples{

# Load data
data("bem_dfmdata")

# Generate model data
model <- create_dfmodel(x = bem_dfmdata, p = 1, n = 1,
                        iterations = 20, burnin = 10)
# Number of iterations and burnin should be much higher.

# Add prior specifications
model <- add_priors(model,
                    lambda = list(vinv = .01),
                    u = list(shape = 5, rate = 4),
                    a = list(vinv = .01),
                    v = list(shape = 5, rate = 4))

# Add initial values
model <- add_initial_values(model)

# Obtain posterior draws
object <- draw_posterior(model)

}
